LLM-1313: Refine tools rendering
