+++ src/extensions/prompt-summary.ts
@@ -121,8 +121,8 @@ export default function promptSummaryExtension(pi: ExtensionAPI) {
 		if (!stats?.tokenUsage) return
 		subagents.input += stats.tokenUsage.input
 		subagents.output += stats.tokenUsage.output
-		// Cache tokens are not reported by the subagent process — subagent cache
-		// usage is not reflected in the summary totals.
+		subagents.cacheRead += stats.tokenUsage.cacheRead
+		subagents.cacheWrite += stats.tokenUsage.cacheWrite
 	})
 
 	pi.on("agent_end", async () => {
